refactor moorhen sequence slider
